Versión 1.1.13. Avance en: * Nuevos campos al enviar los medidores NoRegistrados en el archivo TPL. * En Anomalia, separar los campos en estructura o clase para poder debuggeear. * El recibir en el archivo TPL si intercambiar SerieMedidor vs Código de barras del medidor en la pantalla TomaDeLectura.

+++ app/src/main/java/enruta/sistole_engie/SignaturePadActivity.java
@@ -57,7 +57,7 @@ public class SignaturePadActivity extends Activity {
         ls_nombre= bu_params.getString("ls_nombre");
         temporal = bu_params.getInt("temporal");
         cantidad = bu_params.getInt("cantidad");
-        idLectura= bu_params.getInt("idLectura");
+        idLectura= bu_params.getLong("idLectura");
         thisIsMe= this;
 
         mSignaturePad = (SignaturePad) findViewById(R.id.signature_pad);
